+++ cli/memory_custodian/__init__.py
@@ -1,3 +1,4 @@
 """MemoryCustodian CLI package."""
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
+__protocol_version__ = "0.4"